requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
ase==3.22.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
black==22.3.0
blessed==1.20.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colored==1.4.4
comm==0.2.2
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
e3nn==0.4.4
emmet-core==0.69.5
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
gmpy2==2.1.2
gpustat==1.1.1
grpcio==1.62.1
huggingface-hub==0.21.4
hydra-core==1.3.2
identify==2.5.35
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.3.2
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
latexcodec==2.0.1
lazy_loader==0.3
llvmlite==0.42.0
lmdb==1.4.1
logger==1.4
maggma==0.38.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mongogrant==0.3.3
mongomock==4.0.0
monty==2024.2.26
mp-api==0.37.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
munkres==1.1.4
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.2.1
nodeenv==1.8.0
numba==0.59.0
numpy==1.24.4
omegaconf==2.3.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
orjson==3.9.15
packaging==24.0
palettable==3.3.3
pandas==2.2.1
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.4.0
ply==3.11
pre-commit==2.10.1
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybtex==0.24.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydantic-settings==2.2.1
pydash==7.0.6
Pygments==2.17.2
pymatgen==2023.5.10
pymongo==4.6.2
PyNaCl==1.5.0
pynng==0.8.0
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentinels==1.0.0
sentry-sdk==1.43.0
setproctitle==1.3.3
setuptools==69.2.0
sip==6.7.12
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
spglib==2.3.1
sshtunnel==0.4.0
stack-data==0.6.2
starlette==0.36.3
submitit==1.5.1
sympy==1.12
syrupy==3.0.6
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.0
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.4.12
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typed-ast==1.5.5
typing_extensions==4.10.0
tzdata==2024.1
ukkonen==1.0.1
uncertainties==3.1.7
unicodedata2==15.1.0
urllib3==2.2.1
virtualenv==20.25.1
wandb==0.18.*
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.42.0
yarl==1.9.4
zipp==3.17.0
#
torchdeq
#
sphinx 
nbsphinx 
sphinx-rtd-theme 
pandoc
demjson3
# ocp-models==0.0.3
